Boring type-mismatch fixups (s/unsigned int/NSUInteger/).
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@5972 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Frameworks/EtoileSerialize/Sources/COProxy_NSAttributedString.m
@@ -9,12 +9,12 @@
 #define P3(type, s1, t1, a1, s2, t2, a2, s3, t3, a3) - (type) s1:(t1)a1 s2:(t2)a2 s3:(t3)a3 { return [object s1:a1 s2:a2 s3:a3]; }
 #define P4(type, s1, t1, a1, s2, t2, a2, s3, t3, a3, s4, t4, a4) - (type) s1:(t1)a1 s2:(t2)a2 s3:(t3)a3 s4:(t4)a4 { return [object s1:a1 s2:a2 s3:a3 s4:a4]; }
 @implementation COProxy_NSAttributedString
-P0(unsigned int, length)
+P0(NSUInteger, length)
 P0(NSString*, string)
-P0(unsigned int, _editCount)
+P0(NSUInteger, _editCount)
 P1(BOOL,isEqualToAttributedString,NSAttributedString*, otherString)
-P2(NSDictionary *, attributesAtIndex, unsigned int, index, effectiveRange, NSRangePointer, aRange)
-P3(NSDictionary *, attribute, NSString *, attributeName, atIndex, unsigned int, index, effectiveRange, NSRangePointer, aRange)
-P3(NSDictionary *, attributesAtIndex, unsigned int, index, longestEffectiveRange, NSRangePointer, aRange, inRange,NSRange,rangeLimit)
-P4(NSDictionary *, attribute, NSString *, attributeName, atIndex, unsigned int, index, longestEffectiveRange, NSRangePointer, aRange, inRange,  NSRange,rangeLimit)
+P2(NSDictionary *, attributesAtIndex, NSUInteger, index, effectiveRange, NSRangePointer, aRange)
+P3(NSDictionary *, attribute, NSString *, attributeName, atIndex, NSUInteger, index, effectiveRange, NSRangePointer, aRange)
+P3(NSDictionary *, attributesAtIndex, NSUInteger, index, longestEffectiveRange, NSRangePointer, aRange, inRange,NSRange,rangeLimit)
+P4(NSDictionary *, attribute, NSString *, attributeName, atIndex, NSUInteger, index, longestEffectiveRange, NSRangePointer, aRange, inRange,  NSRange,rangeLimit)
 @end
