+++ Frameworks/EtoileSerialize/Sources/StringMap.h
@@ -4,7 +4,7 @@
  * <p>Simple hash function used to using C strings as keys in <code>NSMapTable</code>s.
  * Simply takes the first few bytes of the string as an int</p>
  */
-unsigned simpleStringHash(NSMapTable *table, const void *anObject);
+NSUInteger simpleStringHash(NSMapTable *table, const void *anObject);
 /**
  * <p>Wrapper around <code>strcmp</code> for use in <code>NSMapTable</code>s.</p>
  */
